Working Sphere To Sphere Physics

+++ LabEngine/Collision.h
@@ -4,12 +4,12 @@ class Collision
 {
 	public:
 		Collision();
-        static bool SphereVersusSphere(const int& x1 ,const int& y1 ,const int& z1 ,const int& size1 ,const int& x2 ,const int& y2 ,const int& z2 ,const int& size2);
-        static bool SphereVersusPoint(const int& sphereX ,const int& sphereY ,const int& sphereZ ,const int& sphereSize ,const int& PointX ,const int& PointY ,const int& PointZ);
-        static bool BoxVersusPoint(const int& PointX ,const int& PointY ,const int& PointZ ,const int& boxMinX ,const int& boxMinY ,const int& boxMinZ ,const int& boxMaxX ,const int& boxMaxY ,const int& boxMaxZ);
-        static bool BoxVersusBox(const int& boxMinX1 ,const int& boxMinY1 ,const int& boxMinZ1 ,const int& boxMaxX1 ,const int& boxMaxY1 ,const int& boxMaxZ1 ,const int& boxMinX2 ,const int& boxMinY2 ,const int& boxMinZ2 ,const int& boxMaxX2 ,const int& boxMaxY2 ,const int& boxMaxZ2);
-        static bool SphereVersusBox(const int& sphereX ,const int& sphereY ,const int& sphereZ ,const int& sphereSize ,const int& boxMinX ,const int& boxMinY ,const int& boxMinZ ,const int& boxMaxX ,const int& boxMaxY ,const int& boxMaxZ);
-        static int Min(const int& i1 , const int& i2);
-        static int Max(const int& i1 ,const int& i2);
+		static bool SphereVersusSphere(const double& x1 ,const double& y1 ,const double& z1 ,const double& size1 ,const double& x2 ,const double& y2 ,const double& z2 ,const double& size2);
+		static bool SphereVersusPoint(const double& sphereX ,const double& sphereY ,const double& sphereZ ,const double& sphereSize ,const double& PointX ,const double& PointY ,const double& PointZ);
+		static bool BoxVersusPoint(const double& PointX ,const double& PointY ,const double& PointZ ,const double& boxMinX ,const double& boxMinY ,const double& boxMinZ ,const double& boxMaxX ,const double& boxMaxY ,const double& boxMaxZ);
+		static bool BoxVersusBox(const double& boxMinX1 ,const double& boxMinY1 ,const double& boxMinZ1 ,const double& boxMaxX1 ,const double& boxMaxY1 ,const double& boxMaxZ1 ,const double& boxMinX2 ,const double& boxMinY2 ,const double& boxMinZ2 ,const double& boxMaxX2 ,const double& boxMaxY2 ,const double& boxMaxZ2);
+		static bool SphereVersusBox(const double& sphereX ,const double& sphereY ,const double& sphereZ ,const double& sphereSize ,const double& boxMinX ,const double& boxMinY ,const double& boxMinZ ,const double& boxMaxX ,const double& boxMaxY ,const double& boxMaxZ);
+		static double Min(const double& i1 , const double& i2);
+		static double Max(const double& i1 ,const double& i2);
 
 };
